hf_deployment/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
python-jose[cryptography]
requests
python-dotenv
faster-whisper
sentence-transformers
google-generativeai
python-Levenshtein
